requirements.txt
flask==3.0.0
python-dotenv==1.0.0
stripe==7.0.0
gunicorn==21.2.0
reportlab==4.0.0

# Phase 2: Uncomment when database and auth are needed
# flask-sqlalchemy==3.1.1
# flask-migrate==4.0.5
# flask-login==0.6.3
# anthropic==0.39.0
